requirements-test.txt
# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0
pytest-asyncio>=0.21.0

# Core dependencies for testing
openai>=1.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
Pillow>=10.0.0

# Development and quality tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Coverage reporting
coverage[toml]>=7.0.0
codecov>=2.1.0